Добавит рефакторинг и упростит код
